skinmatch-backend/requirements.txt
# skinmatch-backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
